Move extensions to FlowCryptCommon
