--- FlowCrypt/Extensions/String+Extension.swift
@@ -1,21 +1,0 @@
-//
-//  String+Extension.swift
-//  FlowCrypt
-//
-//  Created by Roma Sosnovsky on 27/10/21
-//  Copyright © 2017-present FlowCrypt a. s. All rights reserved.
-//
-
-import Foundation
-
-extension String {
-    var isValidEmail: Bool {
-        let emailFormat = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
-        let emailPredicate = NSPredicate(format: "SELF MATCHES %@", emailFormat)
-        return emailPredicate.evaluate(with: self)
-    }
-
-    func capitalizingFirstLetter() -> String {
-        prefix(1).uppercased() + self.lowercased().dropFirst()
-    }
-}
