+++ FlowCrypt/Functionality/Services/ApiCall.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import FlowCryptCommon
 
 enum ApiCall {}
