+++ FlowCrypt/Functionality/Services/GeneralConstants.swift
@@ -13,7 +13,6 @@ enum GeneralConstants {
     }
 
     enum Global {
-        static let generalError = -1
         static let attachmentSizeLimit: Int = 10_000_000
     }
