+++ FlowCrypt/Functionality/Services/Remote Private Key Services/EmailKeyManagerApi.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import FlowCryptCommon
 
 protocol EmailKeyManagerApiType {
     func getPrivateKeys(idToken: String) async throws -> EmailKeyManagerApiResult
