+++ FlowCryptAppTests/Extensions/StringTestExtension.swift
