+++ FlowCryptCommon/Extensions/UIApplicationExtensions.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-extension UIApplication {
+public extension UIApplication {
 
     var keyWindow: UIWindow? {
         // Get connected scenes
